Enable a handler for DM messages
