+++ .vscode/settings.json
@@ -1,3 +1,21 @@
 {
-  "python.formatting.provider": "black"
+  "[python]": {
+    "editor.formatOnSave": true,
+    "editor.defaultFormatter": "ms-python.black-formatter",
+    "editor.codeActionsOnSave": {
+      "source.organizeImports": true
+    },
+  },
+  "python.formatting.provider": "none",
+  "isort.args": [
+    "--profile",
+    "black"
+  ],
+  "python.linting.flake8Enabled": true,
+  "python.linting.lintOnSave": true,
+  "python.linting.pylintEnabled": false,
+  "python.linting.flake8Args": [
+    "--ignore",
+    "E501"
+  ],
 }
